Operator: Namespace objects that runs the components (#5)
Changed the logic of the reconciler to manage a namespace inside the
workspace. Since the namespace does not have a namespace (duh), the
workspace requires a finalizer to clean the namesapce when it gets
deleted.

The OwnerReference also stop working for all the objects since the
workspace will be in another namespace (top level spot-system). The
reason for the workspace to not live under the namespace is because the
process of building images and uploading them to the ECR will require
secrets/configmaps that are not going to be available under that newly
created namespace.

It would be very messy and possibly insecure to copy the secrets/etc over workspace so
it's going to be done at the global namespace instead.

This means a workspace will have half of its lifetime in the global
namespace and the other half (running the actual code) inside its own
managed namespace.