Operator: Implement host/ingress to use Tag and spec  (#11)
Operator: Implement host/ingress to use Tag and spec

There was already a dynamic aspect to the ingress, but it wasn't
complete. Removed all hardcoded values that were related to the ingress
and use Workspace's Tag as well as changed the ServiceSpec for a
NetworkSpec which includes extra data that maps to the Ingress.

Also changed how a workspace lifecycle operates by introducing a new
stage: Networking. The stage is responsible to set the ingress &
services (and eventually the TLS certs) and is run *before* the building
process. This allows k8s to generate all the route and iptable entries
that it needs so that when the deployment happens, the time between the
pod deploying and the connection completed is shorter, which result in a
quicker deployment and makes the DX more enjoyable.