Builder: Implement better support for registries (#23)
Restructuring the code in a way that buildkit is only responsible for
the building of the image. Auth will try to find all the authentication
resources in the pod and store them in a known place.

go-containerregistry is responsible to provide the go structure for the
image index layout (OCI standard) as well as providing the exporting
features so that image index can be uploaded to the registry.

While this is a google's library, the first test I did was with private
ECR which was successful. This increase my confidence that this library
will support all the major registries.

While buildkit can provide all of the functionality, I found that the
 documentation and the configurability to be lackluster. So the code is
 taking an extra step by pushing the image from the daemon to the client
 and then reload that image through go container registry.

 This allows us to have more access to the whole process. There might
 even be a time where I completely remove buildkit for something else.